Fixed NoRotate.
