+++ src/main/java/minegame159/meteorclient/mixininterface/IPlayerMoveC2SPacket.java
@@ -8,8 +8,6 @@ package minegame159.meteorclient.mixininterface;
 public interface IPlayerMoveC2SPacket {
     void setY(double y);
     void setOnGround(boolean onGround);
-    void setYaw(float yaw);
-    void setPitch(float pitch);
     int getTag();
     void setTag(int tag);
 }
