+++ src/main/java/minegame159/meteorclient/mixininterface/IPlayerPositionLookS2CPacket.java
@@ -1,0 +1,6 @@
+package minegame159.meteorclient.mixininterface;
+
+public interface IPlayerPositionLookS2CPacket {
+    void setYaw(float yaw);
+    void setPitch(float pitch);
+}
