+++ src/main/java/minegame159/meteorclient/modules/player/NoRotate.java
@@ -7,12 +7,10 @@ package minegame159.meteorclient.modules.player;
 
 import meteordevelopment.orbit.EventHandler;
 import minegame159.meteorclient.events.packets.PacketEvent;
-import minegame159.meteorclient.mixininterface.IPlayerMoveC2SPacket;
+import minegame159.meteorclient.mixininterface.IPlayerPositionLookS2CPacket;
 import minegame159.meteorclient.modules.Category;
 import minegame159.meteorclient.modules.Module;
-import minegame159.meteorclient.modules.Modules;
-import minegame159.meteorclient.modules.combat.Quiver;
-import net.minecraft.network.packet.c2s.play.PlayerMoveC2SPacket;
+import net.minecraft.network.packet.s2c.play.PlayerPositionLookS2CPacket;
 
 public class NoRotate extends Module {
 
@@ -21,11 +19,10 @@ public class NoRotate extends Module {
     }
 
     @EventHandler
-    private void onSendPacket(PacketEvent.Send event) {
-        if (event.packet instanceof PlayerMoveC2SPacket) {
-            if (Modules.get().get(EXPThrower.class).isActive() || Modules.get().get(Quiver.class).isActive()) return;
-            ((IPlayerMoveC2SPacket) event.packet).setPitch(mc.player.getPitch(0));
-            ((IPlayerMoveC2SPacket) event.packet).setYaw(mc.player.getYaw(1));
+    private void onReceivePacket(PacketEvent.Receive event) {
+        if (event.packet instanceof PlayerPositionLookS2CPacket) {
+            ((IPlayerPositionLookS2CPacket) event.packet).setPitch(mc.player.getPitch(0));
+            ((IPlayerPositionLookS2CPacket) event.packet).setYaw(mc.player.getYaw(1));
         }
     }
 }
