+++ src/main/resources/meteor-client.mixins.json
@@ -106,7 +106,8 @@
     "WorldRendererMixin",
     "TridentItemMixin",
     "ModelPredicateProviderRegistryMixin",
-    "ItemStackAccessor"
+    "ItemStackAccessor",
+    "PlayerPositionLookS2CPacketMixin"
   ],
   "injectors": {
     "defaultRequire": 1
